+++ missiongen/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.8.8"
+__version__ = "1.8.9"
 
 from .recipe import Recipe
 from .builder import StarterBuilder, generate